+++ lib/ts/recipe/dashboard/api/apiKeyProtector.ts
@@ -32,5 +32,7 @@ export default async function apiKeyProtector(
         return true;
     }
 
-    return await apiFunction(apiImplementation, options);
+    const response = await apiFunction(apiImplementation, options);
+    options.res.sendJSONResponse(response);
+    return true;
 }