+++ lib/ts/recipe/dashboard/types.ts
@@ -58,4 +58,30 @@ export type APIInterface = {
     dashboardGET: undefined | ((input: { options: APIOptions; userContext: any }) => Promise<string>);
 };
 
-export type APIFunction = (apiImplementation: APIInterface, options: APIOptions) => Promise<boolean>;
+export type APIFunction = (apiImplementation: APIInterface, options: APIOptions) => Promise<any>;
+
+export type RecipeIdForUser = "emailpassword" | "thirdparty" | "passwordless";
+
+type CommonUserInformation = {
+    id: string;
+    timeJoined: number;
+    firstName: string;
+    lastName: string;
+};
+
+export type EmailPasswordUser = CommonUserInformation & {
+    email: string;
+};
+
+export type ThirdPartyUser = CommonUserInformation & {
+    email: string;
+    thirdParty: {
+        id: string;
+        userId: string;
+    };
+};
+
+export type PasswordlessUser = CommonUserInformation & {
+    email?: string;
+    phone?: string;
+};