+++ lib/ts/recipe/userroles/userRoleClaim.ts
@@ -16,6 +16,7 @@ export class UserRoleClaimClass extends PrimitiveArrayClaim<string> {
                 });
                 return res.roles;
             },
+            defaultMaxAgeInSeconds: 300,
         });
     }
 }